Extract user & comment
