+++ src/Base.php
@@ -31,72 +31,7 @@ abstract class Base {
 		add_action( 'rest_api_init', [ $this, 'init' ] );
 	}
 
-	public function init() {
-		register_rest_field( 'user', 'meta_box', [
-			'get_callback'    => [ $this, 'get_user_meta' ],
-			'update_callback' => [ $this, 'update_user_meta' ],
-		] );
-		register_rest_field( 'comment', 'meta_box', [
-			'get_callback'    => [ $this, 'get_comment_meta' ],
-			'update_callback' => [ $this, 'update_comment_meta' ],
-		] );
-	}
-
-	/**
-	 * Get user meta for the rest API.
-	 *
-	 * @param array $object User object.
-	 *
-	 * @return array
-	 */
-	public function get_user_meta( $object ) {
-		$meta_boxes = rwmb_get_registry( 'meta_box' )->get_by( [ 'object_type' => 'user' ] );
-
-		// Ignore MB User Profile meta boxes.
-		$meta_boxes = array_filter( $meta_boxes, function ( $meta_box ) {
-			return ! in_array( $meta_box->id, [
-				'rwmb-user-register',
-				'rwmb-user-login',
-				'rwmb-user-lost-password',
-				'rwmb-user-reset-password',
-				'rwmb-user-info',
-			], true );
-		} );
-
-		return $this->get_values( $meta_boxes, $object['id'], [ 'object_type' => 'user' ] );
-	}
-
-	/**
-	 * Update user meta for the rest API.
-	 *
-	 * @param string|array $data   User meta values in either JSON or array format.
-	 * @param object       $object User object.
-	 */
-	public function update_user_meta( $data, $object ) {
-		$this->update_values( $data, $object->ID, 'user', 'user' );
-	}
-
-	/**
-	 * Get comment meta for the rest API.
-	 *
-	 * @param array $object Comment object.
-	 *
-	 * @return array
-	 */
-	public function get_comment_meta( $object ) {
-		$meta_boxes = rwmb_get_registry( 'meta_box' )->get_by( [ 'object_type' => 'comment' ] );
-		return $this->get_values( $meta_boxes, $object['id'], [ 'object_type' => 'comment' ] );
-	}
-
-	/**
-	 * Update comment meta for the rest API.
-	 *
-	 * @param string|array $data   Comment meta values in either JSON or array format.
-	 * @param object       $object Comment object.
-	 */
-	public function update_comment_meta( $data, $object ) {
-		$this->update_values( $data, $object->comment_ID, 'comment', 'comment' );
-	}
+	abstract public function init();
 
 	protected function update_values( $data, $object_id, $object_subtype, $object_type ) {
 		$data = is_string( $data ) ? json_decode( $data, true ) : $data;
